fix: rename Scripts and delete unused models and files
