--- Assets/02.Scripts/Objects/PlateDetector.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fbedc91ab60c8fa4bac491804d98fe1a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
