--- Assets/05.Models/Obstacles/aspen-treen-stumps-logs-stack.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: eee875c4fc6a136499fa59757d74a90b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
