--- Assets/05.Models/Obstacles/aspen-treen-stumps-logs-stack/source.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 89af87ff4f970ee47984383e65657bb4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
