--- Assets/05.Models/Obstacles/aspen-treen-stumps-logs-stack/source/stump-stack.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5f589389a1c9cb24086d97d1a449aa0e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
