--- Assets/05.Models/Obstacles/aspen-treen-stumps-logs-stack/textures.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7fd38404288aa184481aee407c5c2eb0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
