--- Assets/05.Models/Obstacles/small-cobblestone-paving-cubes-pile-free.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 10fefadb501b9234d8eb0b68f589e5dc
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
