--- Assets/05.Models/Obstacles/small-cobblestone-paving-cubes-pile-free/source.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6753cc0c3258bac46ac0a7cd41bf4326
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
