--- Assets/05.Models/Obstacles/small-cobblestone-paving-cubes-pile-free/source/small-cubes.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 034bb7806b96151499748c0b2cc4098e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
