--- Assets/05.Models/Obstacles/small-cobblestone-paving-cubes-pile-free/textures.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: eac50693af9dec74684170c415eeb04e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
