--- Assets/05.Models/Obstacles/stylized-rock.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5ccc98e06e1b05345bf0238e04c6c99e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
