--- Assets/05.Models/Obstacles/stylized-rock/source.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 846e7aea18d5bd0428999ed42b4eee83
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
