--- Assets/05.Models/Obstacles/stylized-rock/source/Stylized_Rock.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a8443422992fd0d4284a0a34a219ae86
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
