--- Assets/05.Models/Obstacles/stylized-rock/textures.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d8033e2dd628e064c9345c71b4adc5da
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
