--- Assets/msVFX_Free Smoke Effects Pack.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1f49322f77b6be242872919fa5a2dbd0
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
