+++ Assets/msVFX_Free_Smoke_Effects_Pack/Materials/msVFX_Stylized_Smoke_1_Material.mat.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 3217b570b11b257469c5461f0457e5d4
